# swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The LocalAI Rest API.",
        "title": "LocalAI API",
        "contact": {
            "name": "LocalAI",
            "url": "https://localai.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/mudler/LocalAI/master/LICENSE"
        },
        "version": "2.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/pii/analyze": {
            "post": {
                "description": "Runs the configured PII detectors (NER and/or pattern tiers) over the supplied text and returns the matched entity spans with the policy action that would fire. Detection only — the text is not modified and no block is enforced. Select detectors explicitly via `detectors`, or pass a consuming `model` to use its effective policy: the model's own `pii.detectors`, else the instance-wide `pii_default_detectors`. A model with PII disabled, or enabled with nothing to scan with, is a 400. The raw matched value is never returned; admins may set `reveal:true` for the audit hash prefix.",
                "tags": [
                    "pii"
                ],
                "summary": "Detect PII entities in a string (no mutation).",
                "parameters": [
                    {
                        "description": "text + detector selection",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.PIIAnalyzeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detected entities",
                        "schema": {
                            "$ref": "#/definitions/schema.PIIAnalyzeResponse"
                        }
                    }
                }
            }
        },
        "/api/pii/redact": {
            "post": {
                "description": "Runs the configured PII detectors over the text and applies each detector model's policy: masked spans are replaced with `[REDACTED:\u003cid\u003e]`, allow spans pass through, and a single block action causes a 400 (type `pii_blocked`) carrying the offending entities — the text is never returned in that case. Select detectors via `detectors`, or a consuming `model`'s effective policy (its own `pii.detectors`, else the instance-wide `pii_default_detectors`; PII must be enabled on the model). Records audit events (origin `pii_redact`) visible at /api/pii/events.",
                "tags": [
                    "pii"
                ],
                "summary": "Redact PII in a string by applying the configured policy.",
                "parameters": [
                    {
                        "description": "text + detector selection",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.PIIAnalyzeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Redacted text + entities",
                        "schema": {
                            "$ref": "#/definitions/schema.PIIRedactResponse"
                        }
                    }
                }
            }
        },
        "/api/router/decide": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "router"
                ],
                "summary": "Classify a prompt against a router model's policies (decision oracle)",
                "parameters": [
                    {
                        "description": "decide params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.RouterDecideRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.RouterDecideResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/backend/monitor": {
            "get": {
                "tags": [
                    "monitoring"
                ],
                "summary": "Backend monitor endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the model to monitor",
                        "name": "model",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response",
                        "schema": {
                            "$ref": "#/definitions/proto.StatusResponse"
                        }
                    }
                }
            }
        },
        "/backend/shutdown": {
            "post": {
                "tags": [
                    "monitoring"
                ],
                "summary": "Backend shutdown endpoint",
                "parameters": [
                    {
                        "description": "Backend statistics request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.BackendMonitorRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "proto.MemoryUsageData": {
            "type": "object",
            "properties": {
                "breakdown": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "proto.StatusResponse": {
            "type": "object",
            "properties": {
                "memory": {
                    "$ref": "#/definitions/proto.MemoryUsageData"
                },
                "state": {
                    "$ref": "#/definitions/proto.StatusResponse_State"
                }
            }
        },
        "proto.StatusResponse_State": {
            "type": "integer",
            "format": "int32",
            "enum": [
                0,
                1,
                2,
                -1
            ],
            "x-enum-varnames": [
                "StatusResponse_UNINITIALIZED",
                "StatusResponse_BUSY",
                "StatusResponse_READY",
                "StatusResponse_ERROR"
            ]
        },
...
        "schema.BackendMonitorRequest": {
            "type": "object",
            "properties": {
                "model": {
                    "type": "string"
                }
            }
        },
...
        "schema.PIIAnalyzeRequest": {
            "type": "object",
            "properties": {
                "detectors": {
                    "description": "Detectors names the detector models to run (NER and/or pattern). Takes\nprecedence over Model.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "model": {
                    "description": "Model is a consuming model whose effective PII policy (own\npii.detectors, else the instance default detectors; PII must be\nenabled) is used when Detectors is empty.",
                    "type": "string"
                },
                "reveal": {
                    "description": "Reveal includes the per-entity hash_prefix in the response. Honoured\nonly for admin callers; ignored otherwise. The raw matched value is\nnever returned regardless.",
                    "type": "boolean"
                },
                "text": {
                    "description": "Text is the string to scan. Bounded only by the server's global HTTP\nbody limit.",
                    "type": "string"
                }
            }
        },
        "schema.PIIAnalyzeResponse": {
            "type": "object",
            "properties": {
                "blocked": {
                    "type": "boolean"
                },
                "correlation_id": {
                    "type": "string"
                },
                "entities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.PIIEntity"
                    }
                }
            }
        },
        "schema.PIIEntity": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "end": {
                    "type": "integer"
                },
                "entity_type": {
                    "type": "string"
                },
                "hash_prefix": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "source": {
                    "type": "string"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
        "schema.PIIRedactResponse": {
            "type": "object",
            "properties": {
                "blocked": {
                    "type": "boolean"
                },
                "correlation_id": {
                    "type": "string"
                },
                "entities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.PIIEntity"
                    }
                },
                "masked": {
                    "type": "boolean"
                },
                "redacted_text": {
                    "type": "string"
                }
            }
        },
        "schema.RouterDecideRequest": {
            "type": "object",
            "properties": {
                "input": {
                    "description": "Input is the user-visible prompt text to classify. Required.\nSchema-shape extraction (chat-message concatenation, etc.) is\nthe caller's responsibility — matches the Probe contract used\nby the in-band middleware.",
                    "type": "string"
                },
                "router": {
                    "description": "Router is the name of the router model (a ModelConfig with a\n`router:` block). Required.",
                    "type": "string"
                }
            }
        },
        "schema.RouterDecideResponse": {
            "type": "object",
            "properties": {
                "cache_similarity": {
                    "description": "CacheSimilarity carries the cosine similarity of the cache hit\n(0 when not cached).",
                    "type": "number"
                },
                "cached": {
                    "description": "Cached is true when the decision came from the L2 embedding\ncache rather than a fresh classifier run.",
                    "type": "boolean"
                },
                "candidate": {
                    "description": "Candidate is the model that would be routed to. Empty when no\ncandidate covers Labels AND no fallback is configured.",
                    "type": "string"
                },
                "classifier": {
                    "description": "Classifier is the classifier name that produced the decision\n(e.g. \"score\").",
                    "type": "string"
                },
                "fallback": {
                    "description": "Fallback is true when Candidate is the router's configured\nfallback because no candidate covered Labels. Lets callers\ndistinguish \"matched\" from \"fell back\" without comparing names.",
                    "type": "boolean"
                },
                "labels": {
                    "description": "Labels is the set of active policy labels.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "latency_ms": {
                    "description": "LatencyMs is the classifier's wall-clock cost.",
                    "type": "integer"
                },
                "router": {
                    "description": "Router echoes the requested router model.",
                    "type": "string"
                },
                "score": {
                    "description": "Score is the top label's softmax probability (the\nclassifier-side confidence signal).",
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}